Steering Updated
In this version the steering is updated for improved control. The speed slider switch control is removed and two pirouette-spin switches are added. The diagonal or 'soft' turns operate the turning direction wheels at 50% of the speed of the opposing wheels. The x-axis turns or 'hard' turns 'release' the turning direction wheels.